Move message specific rendering to new file
